extract Complex into its own module
