+++ pgrx-tests/src/tests/mod.rs
@@ -16,6 +16,7 @@ mod attributes_tests;
 mod bgworker_tests;
 mod bytea_tests;
 mod cfg_tests;
+mod complex;
 mod composite_type_tests;
 mod datetime_tests;
 mod default_arg_value_tests;
@@ -63,3 +64,5 @@ mod variadic_tests;
 mod xact_callback_tests;
 mod xid64_tests;
 mod zero_datum_edge_cases;
+
+use complex::Complex;
