+++ pgrx-tests/src/tests/struct_type_tests.rs
@@ -7,96 +7,10 @@
 //LICENSE All rights reserved.
 //LICENSE
 //LICENSE Use of this source code is governed by the MIT license that can be found in the LICENSE file.
-use core::ffi::CStr;
-use pgrx::pgrx_sql_entity_graph::metadata::{
-    ArgumentError, Returns, ReturnsError, SqlMapping, SqlTranslatable,
-};
-use pgrx::prelude::*;
-use pgrx::stringinfo::StringInfo;
-
-use crate::get_named_capture;
-
-#[derive(Debug, Clone, Copy)]
-#[repr(C)]
-pub struct Complex {
-    x: f64,
-    y: f64,
-}
-
-impl Eq for Complex {}
-impl PartialEq for Complex {
-    fn eq(&self, other: &Self) -> bool {
-        self.x == other.x && self.y == other.y
-    }
-}
-
-impl Complex {
-    #[allow(dead_code)]
-    pub fn random() -> PgBox<Complex> {
-        unsafe {
-            let mut c = PgBox::<Complex>::alloc0();
-            c.x = rand::random();
-            c.y = rand::random();
-            c.into_pg_boxed()
-        }
-    }
-}
-
-extension_sql!(
-    r#"CREATE TYPE complex;"#,
-    name = "create_complex_shell_type",
-    creates = [Type(Complex)]
-);
-
-unsafe impl SqlTranslatable for Complex {
-    fn argument_sql() -> Result<SqlMapping, ArgumentError> {
-        Ok(SqlMapping::literal("Complex"))
-    }
-
-    fn return_sql() -> Result<Returns, ReturnsError> {
-        Ok(Returns::One(SqlMapping::literal("Complex")))
-    }
-}
-
-#[pg_extern(immutable)]
-fn complex_in(input: &CStr) -> PgBox<Complex, AllocatedByRust> {
-    let input_as_str = input.to_str().unwrap();
-    let re = regex::Regex::new(
-        r#"(?P<x>[-+]?([0-9]*\.[0-9]+|[0-9]+)),\s*(?P<y>[-+]?([0-9]*\.[0-9]+|[0-9]+))"#,
-    )
-    .unwrap();
-    let x = get_named_capture(&re, "x", input_as_str).unwrap();
-    let y = get_named_capture(&re, "y", input_as_str).unwrap();
-    let mut complex = unsafe { PgBox::<Complex>::alloc() };
-
-    complex.x = str::parse::<f64>(&x).unwrap_or_else(|_| panic!("{x} isn't a f64"));
-    complex.y = str::parse::<f64>(&y).unwrap_or_else(|_| panic!("{y} isn't a f64"));
-
-    complex
-}
-
-#[pg_extern(immutable)]
-fn complex_out(complex: PgBox<Complex>) -> &'static CStr {
-    let mut sb = StringInfo::new();
-    sb.push_str(&format!("{}, {}", complex.x, complex.y));
-    unsafe { sb.leak_cstr() }
-}
-
-extension_sql!(
-    r#"
-CREATE TYPE complex (
-   internallength = 16,
-   input = complex_in,
-   output = complex_out,
-   alignment = double
-);
-"#,
-    name = "create_complex_type",
-    requires = ["create_complex_shell_type", complex_in, complex_out]
-);
+use super::Complex;
 
 #[cfg(any(test, feature = "pg_test"))]
-#[pg_schema]
+#[pgrx::pg_schema]
 mod tests {
     #[allow(unused_imports)]
     use crate as pgrx_tests;
@@ -113,8 +27,8 @@ mod tests {
                 .get_one::<PgBox<Complex>>()?
                 .expect("datum was null");
 
-            assert_eq!(&complex.x, &1.1);
-            assert_eq!(&complex.y, &2.2);
+            assert_eq!(&complex.r, &1.1);
+            assert_eq!(&complex.i, &2.2);
             Ok(())
         })
     }
@@ -135,8 +49,8 @@ mod tests {
                 .get_one::<PgBox<Complex>>()?
                 .expect("datum was null");
 
-            assert_eq!(&complex.x, &1.1);
-            assert_eq!(&complex.y, &2.2);
+            assert_eq!(&complex.r, &1.1);
+            assert_eq!(&complex.i, &2.2);
             Ok(())
         })
     }
@@ -149,8 +63,8 @@ mod tests {
                 SELECT value FROM complex_test ORDER BY id;", None, None)?.first().get_one::<PgBox<Complex>>()
         })?.expect("datum was null");
 
-        assert_eq!(&complex.x, &1.0);
-        assert_eq!(&complex.y, &2.01);
+        assert_eq!(&complex.r, &1.0);
+        assert_eq!(&complex.i, &2.01);
         Ok(())
     }
 }
